pyproject.toml
[tool.poetry]
name = "ai-helpdesk-system"
version = "0.1.0"
description = "An AI-powered help desk system that classifies user requests, retrieves relevant knowledge base information, generates contextual responses using a large language model, and escalates complex issues for human intervention."
authors = ["Oluwaseun Adeyo <oluwaseun@cloudplexo.com>"]
license = "MIT"
readme = "README.md"

# Add this line to point to your src/ directory
packages = [
    { include = "src" }
]

[tool.poetry.dependencies]
python = "^3.10"
numpy = "^2.2.4"
openai = "^1.93.0"
pydantic = "^2.5.0"           # Data validation
uvicorn = "^0.24.0"            # ASGI server for FastAPI
fastapi = "^0.104.1"            # For building the API service
faiss-cpu = "^1.11.0"           # Efficient similarity search for knowledge retrieval
python-dotenv = "^1.1.1"


[tool.poetry.group.dev.dependencies]
black = "^24.8.0"              # Code formatting
flake8 = "^6.0.0"              # Linting
pytest = "^8.3.3"              # Testing framework
pytest-cov = "^4.0.0"
pytest-mock = "^3.14.0"
coverage = "^7.6.1"
pylint = "^3.3.1"
pre-commit = "^3.8.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "--verbose"

[tool.pylint.messages_control]
disable= [
        "bare-except",
        "import-error",
        "duplicate-code",
        "no-else-return",
        "protected-access",
        "too-many-statements",
        "redefined-outer-name",
        "broad-exception-caught",
        "too-few-public-methods",
        "import-outside-toplevel",
        "too-many-return-statements",
        "too-many-instance-attributes"
]

[tool.pylint.MASTER]
ignore-paths = '^Monitoring/.ipynb_checkpoints/.*$|^venv/.*$'

[tool.coverage.run]
branch = true
source = ["src"]

[tool.coverage.report]
show_missing = true
fail_under = 80

[tool.black]
line-length = 83
target-version = ["py38", "py39", "py310", "py311"]